Refactored task processing from ProgressPage down into new ProcessingPage
