+++ Tekapo.sln
@@ -26,6 +26,10 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tekapo.Processing.UnitTests
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tekapo.IntegrationTests", "Tekapo.IntegrationTests\Tekapo.IntegrationTests.csproj", "{BCFB2BFD-DD35-4D45-B94F-1F5ECDFCBD1E}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Application", "Application", "{76340DCA-B140-4B5A-83B7-3DC608D607C6}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{28D99D9B-B097-446F-A134-6E29C150F503}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -63,6 +67,15 @@ Global
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{106EAA95-2C57-4545-935E-7302160C1574} = {76340DCA-B140-4B5A-83B7-3DC608D607C6}
+		{08AC03A4-5267-46F5-989F-B115D9AA3B61} = {76340DCA-B140-4B5A-83B7-3DC608D607C6}
+		{19C1EA4E-AA54-4354-B2FA-D7504F8CF722} = {76340DCA-B140-4B5A-83B7-3DC608D607C6}
+		{0B0E8789-C3CE-475F-B912-62BD156DA605} = {28D99D9B-B097-446F-A134-6E29C150F503}
+		{9D679AD1-0E80-43B5-8D16-2162999C3C79} = {28D99D9B-B097-446F-A134-6E29C150F503}
+		{3D8E335C-FDC0-4087-BD75-1D7072E5A524} = {28D99D9B-B097-446F-A134-6E29C150F503}
+		{BCFB2BFD-DD35-4D45-B94F-1F5ECDFCBD1E} = {28D99D9B-B097-446F-A134-6E29C150F503}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {205BCFA2-5D07-45F1-8D08-52F50D0FE672}
